streaming/requirements.txt
pika==1.3.2
mysql-connector-python==8.1.0